refactor: optimize forum app performance and loading speed(#96451)
