+++ app/[tenant]/[workspace]/(subapps)/forum/common/ui/components/group-controls/group-controls.tsx
@@ -1,0 +1,89 @@
+'use client';
+import {useEffect, useState} from 'react';
+
+// ---- CORE IMPORTS ---- //
+import {i18n} from '@/locale';
+
+// ---- LOCAL IMPORTS ---- //
+import {ForumGroup} from '@/subapps/forum/common/types/forum';
+import {useForum} from '@/subapps/forum/common/ui/context';
+import {
+  GroupActionList,
+  Search as GroupSearch,
+} from '@/subapps/forum/common/ui/components';
+import {GROUPS, MEMBER, OTHER_GROUPS} from '@/subapps/forum/common/constants';
+
+export function GroupControls() {
+  const {memberGroups, nonMemberGroups, user, selectedGroup} = useForum();
+  const [memberGroupList, setMemberGroupList] = useState<ForumGroup[]>(
+    memberGroups || [],
+  );
+  const [nonMemberGroupList, setNonMemberGroupList] = useState<ForumGroup[]>(
+    nonMemberGroups || [],
+  );
+  const userId = user?.id as string;
+  const isLoggedIn = !!user?.id;
+  const [groupSearchValue, setGroupSearchKey] = useState<string>('');
+  const handleGroupSearch = (value: string) => {
+    setGroupSearchKey(value);
+  };
+
+  useEffect(() => {
+    const getGroups = async () => {
+      try {
+        if (groupSearchValue) {
+          const searchKeyLower = groupSearchValue.toLowerCase();
+
+          if (isLoggedIn) {
+            const filteredMemberGroups = memberGroups.filter(
+              (group: ForumGroup) =>
+                group?.forumGroup?.name?.toLowerCase().includes(searchKeyLower),
+            );
+            setMemberGroupList(filteredMemberGroups);
+          }
+          const filteredNonMemberGroups = nonMemberGroups.filter(
+            (group: ForumGroup) =>
+              group?.name?.toLowerCase().includes(searchKeyLower),
+          );
+          setNonMemberGroupList(filteredNonMemberGroups);
+        } else {
+          setMemberGroupList(memberGroups);
+          setNonMemberGroupList(nonMemberGroups);
+        }
+      } catch (err) {
+        console.log(err);
+      }
+    };
+
+    getGroups();
+  }, [isLoggedIn, groupSearchValue, memberGroups, nonMemberGroups]);
+
+  return (
+    <div className="h-fit flex flex-col gap-6 bg-white p-4 rounded-lg">
+      <div>
+        <h1 className="font-semibold text-xl leading-[1.875rem]">
+          {i18n.t(GROUPS)}
+        </h1>
+      </div>
+      <GroupSearch onChange={handleGroupSearch} />
+      {isLoggedIn && (
+        <GroupActionList
+          title={MEMBER}
+          groups={memberGroupList}
+          isMember={true}
+          userId={userId}
+          groupId={selectedGroup?.id}
+        />
+      )}
+      <GroupActionList
+        title={OTHER_GROUPS}
+        groups={nonMemberGroupList}
+        isMember={false}
+        userId={userId}
+        groupId={selectedGroup?.id}
+      />
+    </div>
+  );
+}
+
+export default GroupControls;
