+++ app/[tenant]/[workspace]/(subapps)/forum/common/ui/components/group-controls/index.ts
@@ -1,0 +1,1 @@
+export * from './group-controls';
