+++ app/[tenant]/[workspace]/(subapps)/forum/common/ui/components/hero/index.ts
@@ -1,0 +1,1 @@
+export * from './hero';
