--- app/[tenant]/[workspace]/(subapps)/forum/common/ui/components/home-page/index.ts
@@ -1,1 +1,0 @@
-export * from './home-page';
