+++ app/[tenant]/[workspace]/(subapps)/forum/common/ui/components/index.ts
@@ -2,14 +2,12 @@ export {Search} from './search';
 export {GroupActionList} from './group-action-list';
 export {GroupNotification} from './group-notification';
 export {NavMenu} from './nav-menu';
-export {HomePage} from './home-page';
 export {UploadPost} from './upload-post';
 export {CreatePost} from './create-post';
 export {ImageUploader} from './image-uploader';
 export {FileUploader} from './file-uploader';
 export {FileDesc} from './file-desc';
 export {Tabs} from './tabs';
-export {PostsContent} from './posts-content';
 export {MediaContent} from './media-content';
 export {ThreadList} from './thread-list';
 export {Thread, ThreadHeader, ThreadBody, ThreadFooter} from './thread';
@@ -21,3 +19,5 @@ export {FilePreviewer} from './file-previewer';
 export {ImagePreviewer} from './image-previewer';
 export {ImageViewer} from './image-viewer';
 export {NotificationHeader} from './notification-header';
+export {Hero} from './hero';
+export {GroupControls} from './group-controls';
