+++ app/[tenant]/[workspace]/(subapps)/forum/common/ui/components/nav-menu/nav-menu.tsx
@@ -1,19 +1,17 @@
 'use client';
 
 import React from 'react';
+import Link from 'next/link';
 import {useSession} from 'next-auth/react';
 
 // ---- CORE IMPORTS ---- //
 import {Separator} from '@/ui/components/separator';
+import {SUBAPP_CODES} from '@/constants';
+import {useWorkspace} from '@/app/[tenant]/[workspace]/workspace-context';
 
-export const NavMenu = ({
-  items,
-  onClick,
-}: {
-  items: any;
-  onClick: (link: string) => void;
-}) => {
+export const NavMenu = ({items}: {items: any}) => {
   const {data: session} = useSession();
+  const {workspaceURI} = useWorkspace();
 
   const filteredItems = session
     ? items
@@ -23,12 +21,12 @@ export const NavMenu = ({
     <div className="bg-white flex items-center justify-center gap-5 px-6 py-4">
       {filteredItems.map((item: any, index: number) => (
         <React.Fragment key={item.id}>
-          <div
+          <Link
+            href={`${workspaceURI}/${SUBAPP_CODES.forum}/${item.link}`}
             key={item.id}
-            className="font-medium text-base cursor-pointer"
-            onClick={() => onClick(item.link)}>
+            className="font-medium text-base cursor-pointer">
             {item.name}
-          </div>
+          </Link>
           {index !== filteredItems.length - 1 && (
             <Separator
               className="bg-black w-0.5 h-[1.375rem]"
