--- app/[tenant]/[workspace]/(subapps)/forum/common/ui/components/posts-content/index.ts
@@ -1,1 +1,0 @@
-export * from './posts-content';
