--- app/[tenant]/[workspace]/(subapps)/forum/common/ui/components/posts-content/posts-content.tsx
@@ -1,14 +1,0 @@
-'use client';
-
-// ---- LOCAL IMPORTS ---- //
-import {ThreadList} from '@/subapps/forum/common/ui/components';
-
-export const PostsContent = () => {
-  return (
-    <div className="w-full mt-6">
-      <ThreadList />
-    </div>
-  );
-};
-
-export default PostsContent;
