+++ app/[tenant]/[workspace]/(subapps)/forum/common/ui/components/thread-list/thread-list.tsx
@@ -11,13 +11,17 @@ import {SORT_BY_OPTIONS} from '@/comments';
 
 // ---- LOCAL IMPORTS ---- //
 import {InfiniteScroll} from '@/subapps/forum/common/ui/components';
-import {useForum} from '@/subapps/forum/common/ui/context';
+import {Post} from '@/subapps/forum/common/types/forum';
 
-export const ThreadList = () => {
+export const ThreadList = ({
+  posts,
+  pageInfo,
+}: {
+  posts: Post[];
+  pageInfo: any;
+}) => {
   const {update, searchParams} = useSearchParams();
   const sort = searchParams.get('sort') ?? 'new';
-
-  const {posts, pageInfo} = useForum();
 
   const handleSortBy = (value: any) => {
     if (!value) {
